+++ R/text-to-speech.R
@@ -12,7 +12,8 @@
 #' @param pitch Speaking pitch between \code{-20.0} and \code{20.0} in semitones.
 #' @param volumeGainDb Volumne gain in dB
 #' @param sampleRateHertz Sample rate for returned audio
-#' @param inputType Choose between \code{text} (the default) or SSML markup. The \code{input} text must be SSML markup if you choose \code{ssml}.
+#' @param inputType Choose between \code{text} (the default) or SSML markup. The \code{input} text must be SSML markup if you choose \code{ssml}
+#' @param effectsProfileIds Optional. An identifier which selects 'audio effects' profiles that are applied on (post synthesized) text to speech. Effects are applied on top of each other in the order they are given
 #'
 #' @details
 #'
@@ -24,13 +25,15 @@
 #'
 #' To use Speech Synthesis Markup Language (SSML) select \code{inputType=ssml} - more details on using this to insert pauses, sounds and breaks in your audio can be found here: \url{https://cloud.google.com/text-to-speech/docs/ssml}
 #'
+#' To use audio profiles, supply a character vector of the available audio profiles listed here: \url{https://cloud.google.com/text-to-speech/docs/audio-profiles} - the audio profiles are applied in the order given.  For instance \code{effectsProfileIds="wearable-class-device"} will optimise output for smart watches, \code{effectsProfileIds=c("wearable-class-device","telephony-class-application")} will apply sound filters optimised for smart watches, then telephonic devices.
+#'
 #' @seealso \url{https://cloud.google.com/text-to-speech/docs/}
 #'
 #' @return The file output name you supplied as \code{output}
 #' @examples
 #'
 #' \dontrun{
-#'
+#' library(magrittr)
 #' gl_talk("The rain in spain falls mainly in the plain",
 #'         output = "output.wav")
 #'
@@ -41,6 +44,10 @@
 #'   standard <break time=\"1s\"/>is defined by the
 #'   <sub alias=\"World Wide Web Consortium\">W3C</sub>.</speak>',
 #'   inputType =  "ssml")
+#'
+#' # using effects profiles
+#' gl_talk("This sounds great on headphones",
+#'         effectsProfileIds = "headphone-class-device")
 #'
 #' }
 #'
@@ -57,7 +64,8 @@ gl_talk <- function(input,
                     pitch = 0,
                     volumeGainDb = 0,
                     sampleRateHertz = NULL,
-                    inputType = c("text","ssml")){
+                    inputType = c("text","ssml"),
+                    effectsProfileIds = NULL){
 
   inputType     <- match.arg(inputType)
   gender        <- match.arg(gender)
@@ -108,6 +116,11 @@ gl_talk <- function(input,
     )
   }
 
+  if(!is.null(effectsProfileIds)){
+    assert_that(is.character(effectsProfileIds))
+    effectsProfileIds <- as.list(effectsProfileIds)
+  }
+
   body <- list(
     input = the_input,
     voice = list(
@@ -120,7 +133,8 @@ gl_talk <- function(input,
       speakingRate = speakingRate,
       pitch = pitch,
       volumeGainDb = volumeGainDb,
-      sampleRateHertz = sampleRateHertz
+      sampleRateHertz = sampleRateHertz,
+      effectsProfileId = effectsProfileIds
     )
   )